Aufgabe 1 // Datentypenzuordnung

+++ src/aufgabe1/Aufgabe1.java
@@ -11,7 +11,7 @@ public class Aufgabe1 {
     double b = 5.0;
     // = 5.0
 
-    // c) 4L / String
+    // c) 4L / long
     long c = 4;
     // = 4
 
@@ -19,31 +19,31 @@ public class Aufgabe1 {
     int d = 3 + 9;
     // = 12
 
-    // e) 2 + 6.0
+    // e) 2 + 6.0 / double
     double e = 2 + 6.0;
     // = 8
 
-    // f) 5 + 3.0f
-    double f = 5 + 3.0;
+    // f) 5 + 3.0f / float
+    float f = 5 + 3.0;
     // = 8
 
-    // g) 3d + 1.0f
+    // g) 3d + 1.0f / double
     double g = 3 + 1.0;
     // = 4
 
-    // h) 6L * 3
+    // h) 6L * 3 / long
     long h= 6 * 3;
     // = 18
 
     // i) 7 / 2
-    double i = 7 / 2;
+    double i = 7 / 2; // !!! int i = 7/2 ist richtig dementsprechend lösung = 3
     // = 3.5
 
-    // j) 9 / 2.0
+    // j) 9 / 2.0 / double
     double j = 9 / 2.0;
     // = 4.5
 
-    // k) int i = 2; i--;
+    // k) int i = 2; i--; / int
     int x = 2;
     int k = x--;
     // = 1
